backend/requirements.txt
Flask
flask_cors
torch
torchvision
Pillow
numpy
open-cv-python